+++ compiler/test/data/idl/callbackInterface.d.kt
@@ -1,0 +1,5 @@
+
+external interface EventListener {
+    fun handleEvent(event: Event)
+}
+external abstract class Event