+++ translator-string/src/org/jetbrains/dukat/translatorString/StringTranslator.kt
@@ -35,11 +35,7 @@ const val FORMAT_TAB = "    "
 
 private fun String?.translateMeta(): String {
     return if (this != null) {
-        if (this.startsWith("=")) {
-            " /*${this}*/"
-        } else {
-            " /* ${this} */"
-        }
+        " /* ${this} */"
     } else {
         ""
     }
@@ -77,7 +73,12 @@ fun TypeModel.translate(): String {
             val res = mutableListOf("(")
             val paramsList = mutableListOf<String>()
             for (param in parameters) {
-                val paramSerialized = param.name + ": " + param.type.translate() + param.type.translateMeta()
+                val paramNameSerialized = if (param.name != "") {
+                    param.name + ": "
+                } else {
+                    ""
+                }
+                val paramSerialized = paramNameSerialized + param.type.translate() + param.type.translateMeta()
                 paramsList.add(paramSerialized)
             }
             res.add(paramsList.joinToString(", ") + ")")