+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+#### [0.0.13] - 24'July, 2019
+ - [idl] support for translating dictionaries, nullable types, array types and parameter types
+ - [idl] support "implements" statement
+ - redundant Unit return type in not printed anymore
+
 #### [0.0.12] - 19'July, 2019
  - package name (if found) is added as a prefix to the name of the translated entity
  - [#73 - Union type unrolling is broken when it's aliased in a referenced file](https://github.com/Kotlin/dukat/issues/73)